% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preModel.R
\name{bin_analysis}
\alias{bin_analysis}
\title{Buildng plots for cheak binning}
\usage{
bin_analysis(variable_fc_bin, target_calc_bin, bin_data, IsOpen = T)
}
\arguments{
\item{variable_fc_bin}{Number of column first binned variable in dataset}

\item{target_calc_bin}{Number of column target`s variable}

\item{bin_data}{Dataset with binned data}

\item{isOpen}{= T Open html file automatically}
}
\value{
A dataframe with binned variables
}
\description{
\code{bin_analysis} will buildig plots and analysis
}
