% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preModel.R
\name{binning}
\alias{binning}
\title{Add binned variables to the dataset}
\usage{
binning(df, variable_fc, target_calc)
}
\arguments{
\item{df}{A dataframe.}

\item{variable_fc}{Number of first column of variable for binning}

\item{target_calc}{Number of column target`s variable}
}
\value{
A dataframe with binned variables
}
\description{
\code{binning} binning numeric variables
}
