% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preModel.R
\name{delVar}
\alias{delVar}
\title{Delete Variables Based on Single-Value Percent}
\usage{
delVar(df, percent = 0.9, exclude = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{percent}{The given threshold, default \code{0.9}.}

\item{exclude}{Vector of column names or numbers to exclude, default \code{NULL}.}
}
\value{
A dataframe after deleted
}
\description{
\code{delVar} will delete variables whose single-value percent is more than or equal to the given threshold, then return a new dataframe.
}
