% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preModel.R
\name{maxPercVar}
\alias{maxPercVar}
\title{Get the Max Percents of All Variable's of one value}
\usage{
maxPercVar(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A vector
}
\description{
\code{maxPercVar} will return the max percent vector of every variable's
}
