% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preModel.R
\name{maxPercVar_x}
\alias{maxPercVar_x}
\title{Get the Max Percent of the Variable's values}
\usage{
maxPercVar_x(df, x_nm)
}
\arguments{
\item{df}{A dataframe.}

\item{x_nm}{Name of variable.}
}
\value{
A numeric
}
\description{
Get the Max Percent of the Variable's values
}
